+++ Source/SwiftLintFramework/Rules/MultilineArgumentsConfiguration.swift
@@ -3,7 +3,7 @@ import Foundation
 private enum ConfigurationKey: String {
     case severity = "severity"
     case firstArgumentLocation = "first_argument_location"
-    case allowFirstClosureOnSameLine = "allow_first_closure_on_same_line"
+    case onlyEnforceAfterFirstClosureOnFirstLine = "only_enforce_after_first_closure_on_first_line"
 }
 
 public struct MultilineArgumentsConfiguration: RuleConfiguration, Equatable {
@@ -25,12 +25,12 @@ public struct MultilineArgumentsConfiguration: RuleConfiguration, Equatable {
 
     private(set) var severityConfiguration = SeverityConfiguration(.warning)
     private(set) var firstArgumentLocation = FirstArgumentLocation.anyLine
-    private(set) var allowFirstClosureOnSameLine = false
+    private(set) var onlyEnforceAfterFirstClosureOnFirstLine = false
 
     public var consoleDescription: String {
         return severityConfiguration.consoleDescription +
             ", \(ConfigurationKey.firstArgumentLocation.rawValue): \(firstArgumentLocation.rawValue)" +
-            ", \(ConfigurationKey.allowFirstClosureOnSameLine.rawValue): \(allowFirstClosureOnSameLine)"
+            ", \(ConfigurationKey.onlyEnforceAfterFirstClosureOnFirstLine.rawValue): \(onlyEnforceAfterFirstClosureOnFirstLine)"
     }
 
     public mutating func apply(configuration: Any) throws {
@@ -50,8 +50,8 @@ public struct MultilineArgumentsConfiguration: RuleConfiguration, Equatable {
                 try firstArgumentLocation = FirstArgumentLocation(value: value)
             case (.severity, let stringValue as String):
                 try severityConfiguration.apply(configuration: stringValue)
-            case (.allowFirstClosureOnSameLine, let boolValue as Bool):
-                allowFirstClosureOnSameLine = boolValue
+            case (.onlyEnforceAfterFirstClosureOnFirstLine, let boolValue as Bool):
+                onlyEnforceAfterFirstClosureOnFirstLine = boolValue
             default:
                 throw error
             }